+++ db/rocksdb/db_test.go
@@ -1,6 +1,7 @@
-package badgerdb
+package rocksdb
 
 import (
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -38,4 +39,25 @@ func TestRevert(t *testing.T) {
 
 func TestReloadDB(t *testing.T) {
 	dbtest.DoTestReloadDB(t, load)
+}
+
+// Test that the DB can be reloaded after a failed Revert
+func TestRevertRecovery(t *testing.T) {
+	dir := t.TempDir()
+	db, err := NewDB(dir)
+	require.NoError(t, err)
+	_, err = db.SaveNextVersion()
+	require.NoError(t, err)
+	txn := db.Writer()
+	require.NoError(t, txn.Set([]byte{1}, []byte{1}))
+	require.NoError(t, txn.Set([]byte{2}, []byte{2}))
+	require.NoError(t, txn.Commit())
+
+	// make checkpoints dir temporarily unreadable to trigger an error
+	require.NoError(t, os.Chmod(db.checkpointsDir(), 0000))
+	require.Error(t, db.Revert())
+
+	require.NoError(t, os.Chmod(db.checkpointsDir(), 0755))
+	db, err = NewDB(dir)
+	require.NoError(t, err)
 }