+++ x/auth/tx/amino_aux_test.go
@@ -113,7 +113,6 @@ func TestAminoAuxHandler_DefaultMode(t *testing.T) {
 func TestAminoAuxModeHandler_nonDIRECT_MODE(t *testing.T) {
 	invalidModes := []signingtypes.SignMode{
 		signingtypes.SignMode_SIGN_MODE_DIRECT,
-		signingtypes.SignMode_SIGN_MODE_DIRECT_JSON,
 		signingtypes.SignMode_SIGN_MODE_TEXTUAL,
 		signingtypes.SignMode_SIGN_MODE_LEGACY_AMINO_JSON,
 		signingtypes.SignMode_SIGN_MODE_UNSPECIFIED,